Scraper algorithm now works fully with all options
The algorithm is not very effecient as it is in early stages of development and requires tweaking. However, it works as long as the endpoints work.

Additonally, every request consumes an extra API request from STRATZ in order to keep track of the service and provide user feedback if something is wrong with the API.

Matchups now takes list objects rather than strings.

Removed reduduant jquery statements.